[Arkos] Added SFX support for AKG and AKM replayers + Optimized/fixed argument usage + Updated sample program to showcase SFX usage [Pletter] Optimized interruption disabling usage [Content] Update font, cursor and precomputed math tables data

+++ engine/src/arkos/aky_player.h
@@ -18,7 +18,7 @@
 // Paramaters:
 //   data	- Pointer to the music data
 //   num	- Music number
-void AKY_Init(const void* data, u16 num);
+void AKY_Init(const void* data, u8 num);
 
 // Function: AKY_Stop
 // Stop music playback
